requirements.txt
# Core FastAPI dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.15
alembic>=1.10.4
psycopg2-binary>=2.9.6

# Authentication and security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pyotp>=2.8.0
python-multipart>=0.0.6

# Storage
boto3>=1.26.133

# PDF generation
pyppeteer>=1.0.2
PyPDF2>=3.0.1
Pillow>=9.5.0

# HTTP client
httpx>=0.24.0

# Email
python-dotenv>=1.0.0

# Utilities
slugify>=0.0.1